requirements.txt
pandas
numpy
beautifulsoup4
requests
qdrant-client
Pillow
ipython
google-cloud-aiplatform
streamlit